% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{MSfrags}
\alias{MSfrags}
\title{Title}
\format{
The peptide sequence length that is a character vector.
}
\source{
data-raw/data.R
}
\usage{
MSfrags
}
\description{
Fragment with protein modification that we expect
to find in the CCR5 gene from performing mass sepc.
}
\keyword{datasets}
