% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseMSstrings.R
\name{parse_mods}
\alias{parse_mods}
\title{This Function takes strings from Spectronaut output and, pull metadata protein
modifications, and displays the data in a nice format.}
\usage{
parse_mods(seqs, format = "Spectronaut")
}
\arguments{
\item{seqs}{Character vector, peptide sequence strings from Spectronaut output}

\item{format}{Character value, format of the input strings}
}
\value{
A \code{protein_mods} object, which is a data.frame with the amino acid sequences and modifications
}
\description{
This Function takes strings from Spectronaut output and, pull metadata protein
modifications, and displays the data in a nice format.
}
